Add routers and redis manager
